requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
async-timeout==5.0.1
attrs==26.1.0
blinker==1.9.0
bottle==0.13.4
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.1.8
edge-tts==7.2.8
Flask==3.1.3
frozenlist==1.8.0
idna==3.13
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
langid==1.1.6
MarkupSafe==3.0.3
multidict==6.7.1
Nuitka==4.0.8
numpy==2.4.4
ordered-set==4.1.0
packaging==26.2
propcache==0.4.1
proxy_tools==0.1.0
pyobjc-core==12.1; platform_system=="Darwin"
pyobjc-framework-Cocoa==12.1; platform_system=="Darwin"
pyobjc-framework-Quartz==12.1; platform_system=="Darwin"
pyobjc-framework-Security==12.1; platform_system=="Darwin"
pyobjc-framework-UniformTypeIdentifiers==12.1; platform_system=="Darwin"
pyobjc-framework-WebKit==12.1; platform_system=="Darwin"
pywebview==6.2.1
requests==2.32.5
tabulate==0.9.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
yarl==1.22.0
zipp==3.23.1
zstandard==0.25.0
